Implemented tests for search use case
