+++ Test/src/UseCaseTests.java
@@ -1,5 +1,6 @@
-import entity.CommonProductFactory;
+import entity.CommonProduct;
 import entity.Product;
+import entity.CommonProductFactory;
 import entity.ProductFactory;
 import interface_adapter.Search.SearchController;
 import interface_adapter.Search.SearchPresenter;
@@ -14,13 +15,36 @@ import use_case.search.SearchInputBoundary;
 import use_case.search.SearchInteractor;
 import use_case.search.SearchOutputBoundary;
 import use_case.shopping_list.InMemoryShoppingListDataAccess;
-import use_case.shopping_list.add.AddInteractor;
-import use_case.shopping_list.add.AddOutputBoundary;
-import view.SearchView;
+import use_case.shopping_list.add.*;
+import app.SearchUseCaseFactory;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import java.util.List;
 
-public class SearchTest {
+import static org.junit.jupiter.api.Assertions.*;
+public class UseCaseTests {
+
+    @Test
+    public void testAddProductToShoppingList_Success() {
+        InMemoryShoppingListDataAccess mockDataAccess = new InMemoryShoppingListDataAccess();
+        ProductFactory productFactory = new CommonProductFactory();
+        ViewManagerModel viewManagerModel = new ViewManagerModel();
+        ShoppingListViewModel shoppingListViewModel = new ShoppingListViewModel();
+        Product sampleProduct = productFactory.create("Soccer Ball", 10.99f, "Nike", "Nice Ball", "", 0.4f);
+        AddOutputBoundary mockAddOutputBoundary = new ShoppingListPresenter(viewManagerModel, shoppingListViewModel);
+        AddInteractor addInteractor = new AddInteractor(mockDataAccess, mockAddOutputBoundary);
+
+
+        AddController addController = new AddController(addInteractor);
+        addController.execute(sampleProduct);
+
+        assertTrue(mockDataAccess.getShoppingList().contains(sampleProduct));
+        assertTrue(shoppingListViewModel.getState().getProductList().contains(sampleProduct));
+    }
+    @Test
+    public void testAddProductToShoppingList_Failure() {
+        // Doesn't fail because there is an explicit if statement to check if the product is null
+
+    }
     @Test
     public void testSearchProduct_Success() {
         ViewManagerModel viewManagerModel = new ViewManagerModel();
